+++ spec/fixtures/cookbooks/oneview_test_api600_c7000/metadata.rb
@@ -1,0 +1,7 @@
+name             'oneview_test_api600_c7000'
+maintainer       'none'
+license          'All rights reserved'
+description      'A test cookbook for the oneview cookbook API600::C7000 module'
+version          '0.1.0'
+
+depends          'oneview'