+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 3.2.1(unreleased)
+### Notes
+- This release adds support to Oneview Rest API version 600, 800, 1000, and 1200 for C7000 and Synergy to the features Server Certificate.
+
+### New features and resources
+  - Added support for following resources
+    - oneview_server_certificate
 
 ## Bugs & Enhancements:
 - [#419](https://github.com/HewlettPackard/oneview-chef/issues/419) Dockerfile is failing because of insufficient libraries.